database/requirements.txt
firebase_admin
google-generativeai
python-dateutil